+++ app/about.php
@@ -14,28 +14,13 @@
 
 <html>
   <head>
-    <title>Naomi R. Patton | About</title>
+    <title>Naomi R. Patton</title>
     <meta charset="utf-8">
-    
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
-
-    <script type="text/javascript" src="/js/main.js"></script>
-
-    <script type="text/javascript" src="http://use.typekit.com/czg3tgq.js"></script>
-    <script type="text/javascript">
-      try {
-        Typekit.load({
-          active: function() {
-            bW.go();
-          }
-        })
-      }
-      catch (e) {}
-    </script>
     
     <script type="text/javascript" src="js/jquery.js"></script>
     <script type="text/javascript" src="js/jquery.lightbox-0.5.js"></script>
     
+    <?php include('typefaces.php'); ?>
     <link rel="stylesheet" href="/css/main.css" media="screen" />
     
   </head>
@@ -43,11 +28,9 @@
     <div id="membrane">
     
       <div id="masthead">
-        <h1><em>Naomi</em> R. Patton</h1>
+        <h1><a href="index.php" title="Home page"><em>Naomi</em> R. Patton</a></h1>
         <ul id="nav">
-          <?php
-            require 'nav.php';
-          ?>
+          <?php include('nav.php'); ?>
         </ul>
       </div>
       
@@ -55,15 +38,16 @@
       
         <div id="about-block">
           <h2><a name="nrphome" id="about">About Naomi</a></h2>
-          <p class="summary-head">Naomi Patton oversees strategic communications planning and development for all concert/entertainment events and corporate communications for Palace Sports & Entertainment properties: The Palace of Auburn Hills, DTE Energy Music Theatre, Meadow Brook Music Festival and the Detroit Pistons.</p>
-          <p class="about">Patton previously served as Detroit Mayor Dave Bing’s press secretary as he worked to fiscally stabilize the city and reform and restructure city government. In addition her role as speechwriter, she was responsible for developing and managing strategic short- and long-term messaging across platforms and media relations.</p>
-         <p class="about">Prior to joining the Bing administration in 2012, Patton was an enterprise and general assignment reporter for the Detroit Free Press. She covered the Wayne County suburbs and Detroit’s City Hall beat at the height of the text message scandal that forced former Mayor Kwame Kilpatrick to resign, and was part of the team that earned the 2009 Pulitzer Prize for Local Reporting.</p>
-         <p class="about">Before heading to the Free Press in April 2003, Patton worked for the Wisconsin State Journal in Madison where she was a general assignment reporter, covering police, city government, and features. And she began her newspaper career at the Milwaukee Community Journal, Wisconsin's largest African-American newspaper, as a reporter and managing editor of the paper's WEEKEND section. Patton also worked in corporate and public affairs communications in Washington, D.C.</p>
+          <p class="about">Naomi R. Patton is an award-winning journalist who worked for nearly 20 years covering enterprise, government, police and courts with the Milwaukee Community Journal, the Wisconsin State Journal and the Detroit Free Press, where she was member of the team which earned the 2009 Pulitzer Prize for Local Reporting.</p>
+          <p class="about">Since leaving journalism to work as press secretary for Detroit Mayor Dave Bing, Naomi has worked in the public and private sectors in government, the NBA and for two global public relations agencies. After leaving the mayor’s office, she helped lead corporate and entertainment public relations and community relations communications for Palace Sports & Entertainment and the Detroit Pistons.</p>
+          <p class="about">She transitioned into the agency world, continuing to write, develop and produce content, while leading and managing strategic communications for foundation, C-Suite, CSR, nonprofit and corporate clients at Weber Shandwick. At Ruder Finn, Naomi focused on issues and reputation management, where client engagements included emergency management, data breaches and thought leadership.</p>
         </div>
 
       </div>
       
     </div>
   </body>
+
+  <script type="text/javascript" src="/js/main.js"></script>
 
 </html>