+++ app/index.php
@@ -17,25 +17,10 @@
     <title>Naomi R. Patton</title>
     <meta charset="utf-8">
     
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
-
-    <script type="text/javascript" src="/js/main.js"></script>
-
-    <script type="text/javascript" src="http://use.typekit.com/czg3tgq.js"></script>
-    <script type="text/javascript">
-      try {
-        Typekit.load({
-          active: function() {
-            bW.go();
-          }
-        })
-      }
-      catch (e) {}
-    </script>
-    
     <script type="text/javascript" src="js/jquery.js"></script>
     <script type="text/javascript" src="js/jquery.lightbox-0.5.js"></script>
     
+    <?php include('typefaces.php'); ?>
     <link rel="stylesheet" href="/css/main.css" media="screen" />
     
   </head>
@@ -43,21 +28,21 @@
     <div id="membrane">
     
       <div id="masthead">
-        <h1><em>Naomi</em> R. Patton</h1>
+        <h1><a href="index.php" title="Home page"><em>Naomi</em> R. Patton</a></h1>
         <ul id="nav">
-          <?php require 'nav.php' ?>
+          <?php include('nav.php'); ?>
         </ul>
       </div>
       
       <div id="main">
       
         <div id="home-block">
-          <h2><a name="nrphome" id="home">Home</a></h2>
-          <p class="summary-head">In digital and traditional forms, from corporate and public affairs communications to small papers and metropolitan dailies, I share information that puts you in the room, that compels people to act.</p>
+          <p class="summary-first">I’m more than a writer, I’m a communicator. In digital and traditional forms, from corporate, public affairs and social sector communications to weekly newspapers and metropolitan dailies, I tell and elevate the story.</p>
           
           <img src="/photos/naomi_home_portrait3.jpg" alt="Naomi R. Patton has worked in journalism and communications for nearly 20 years." id="home-portrait" />
-          <p class="summary">I chronicled insurance redlining in Milwaukee's central city. I helped develop national campaigns and website content for Washington, D.C.-based professional and trade associations. I worked on the editing team for a book on the history of the Harley-Davidson company. I stood in Metro Airport, December 25, 2009, for hours until federal agents left after apprehending a would-be terrorist who tried to blow up a commercial flight with explosives in his underwear.</p>
-          <p class="summary" style="margin-top: 20px;">Spokesperson. Reporter. Content Producer.  The titles change. The platforms evolve. The storytelling is constant.</p>
+          <p class="summary">I’ve developed messaging and communication campaigns, highlighting Detroit Mayor Dave Bing’s policy agenda and achievements. And I've managed integrated marketing campaigns as the strategic corporate and entertainment communications lead for the parent company of an NBA team.</p>
+          <p class="summary">Working extensively as counsel to public relations agency clients across philanthropy, public affairs, corporate communications, thought leadership and issues and reputation management, I’ve approached each as an opportunity to invest in their people and purpose, to successfully tell their story.</p>
+          <p class="summary" style="margin-top: 20px;">Writer. Speechwriter. Public Relations Consultant. Content Strategist. The titles change. The platforms evolve. The storytelling is constant.</p>
         </div>
 
       </div>
@@ -65,4 +50,5 @@
     </div>
   </body>
 
+  <script type="text/javascript" src="/js/main.js"></script>
 </html>